+++ pkg/image/context.go
@@ -41,14 +41,16 @@ type ArtifactSources struct {
 	} `yaml:"endpoint-copier-operator"`
 	Kubernetes struct {
 		K3s struct {
-			SELinuxPackage    string `yaml:"selinuxPackage"`
-			SELinuxRepository string `yaml:"selinuxRepository"`
-			ReleaseURL        string `yaml:"releaseURL"`
+			SELinuxPackage            string `yaml:"selinuxPackage"`
+			SELinuxRepository         string `yaml:"selinuxRepository"`
+			SELinuxRepositoryPriority int    `yaml:"selinuxRepositoryPriority"`
+			ReleaseURL                string `yaml:"releaseURL"`
 		} `yaml:"k3s"`
 		Rke2 struct {
-			SELinuxPackage    string `yaml:"selinuxPackage"`
-			SELinuxRepository string `yaml:"selinuxRepository"`
-			ReleaseURL        string `yaml:"releaseURL"`
+			SELinuxPackage            string `yaml:"selinuxPackage"`
+			SELinuxRepository         string `yaml:"selinuxRepository"`
+			SELinuxRepositoryPriority int    `yaml:"selinuxRepositoryPriority"`
+			ReleaseURL                string `yaml:"releaseURL"`
 		} `yaml:"rke2"`
 	} `yaml:"kubernetes"`
 }