+++ RELEASE_NOTES.md
@@ -9,6 +9,8 @@
 * Added ability to generate independent combustion drive
 * Added warning about providing container images with index digests
 * Updated documentation to specify that `os-files` should be used for `HelmChartConfigs` instead of using the respective `/kubernetes/manifests` directory
+* Dependency upgrades
+  * Embedded registry is now utilizing Hauler v1.2.5 (upgraded from v1.2.1)
 
 ## API